Add CMakeFiles to build cephes and install doc
The build only links functions for double precision, because the other
precisions cause some name clashes. It also ignores the file for rational
polynomial functions, whose functions' names clash with the simpler polynomials.

When files exist in exact duplicate in several directories,
cmath is preferred, then the folder, then misc

Links to the documentation is added to the local torch documentation.