Fix problem in calling row vectors
The problem was in the expand: it expanded parameters who had one single
row, thus expanding row vectors of the right size.
Instead, it now expands parameters which have a single element. We loose
the ability to expand parameters which should have expanded row vectors
(with one row per output), but I do not think this case occurs in
Cephes.

TODO: we should write more extensive testing of the automatic wrapper
for all the possible input/output. Already pushing this fix.